+++ CHANGELOG.markdown
@@ -1,3 +1,7 @@
+next
+----
+* Add a `Generic` default implementation for `Wrapped`
+
 4.15.1
 ----
 * Restore the `generic` and `generic1` functions in `GHC.Generics.Lens`